requirements.txt

aioredis>=1.3.1
amqp>=5.0.6
asgiref>=3.3.4
async-timeout>=3.0.1
attrs>=21.2.0
autobahn>=21.3.1
Automat>=20.2.0
billiard>=3.6.4.0
celery>=5.1.0
certifi>=2021.5.30
cffi>=1.14.5
channels>=3.0.3
channels-redis>=3.2.0
chardet>=4.0.0
click>=7.1.2
click-didyoumean>=0.0.3
click-plugins>=1.1.1
click-repl>=0.2.0
constantly>=15.1.0
cryptography>=3.4.7
daphne>=3.0.2
Django>=3.2.5
django-appconf>=1.0.4
django-celery-results>=2.1.0
django-cors-headers>=3.8.0
django-debug-toolbar>=3.2.1
django-filter>=2.4.0
django-redis>=5.0.0
django-simple-history>=3.0.0
django-taggit>=1.4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.1
gunicorn>=20.1.0
hiredis>=2.0.0
hyperlink>=21.0.0
idna>=2.10
incremental>=21.3.0
kombu>=5.1.0
msgpack>=1.0.2
mysqlclient>=2.0.3
phonenumbers>=8.12.25
Pillow>=8.2.0
prompt-toolkit>=3.0.18
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.20
PyJWT>=2.1.0
pymemcache>=3.4.4
pyOpenSSL>=20.0.1
pyotp>=2.6.0
python-dateutil>=2.8.1
python-memcached>=1.59
pytz>=2021.1
qrcode>=7.2
redis>=3.5.3
requests>=2.25.1
sentry-sdk>=1.1.0
service-identity>=21.1.0
six>=1.16.0
sqlparse>=0.4.1
Twisted>=21.2.0
txaio>=21.2.1
urllib3>=1.26.5
vine>=5.0.0
wcwidth>=0.2.5
whitenoise>=5.2.0
zope.interface>=5.4.0
